requirements.txt
Pillow==9.5.0
Pillow==10.2.0
Requests==2.31.0
streamlit==1.31.1
torch==2.2.1
torchvision==0.17.1
